+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to WWPTools will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.3] - 2026-01-09
+
+### Fixed
+- Installer License Agreement dialog UI errors (first active control and tab order)
+- MSI build step now patches dialog metadata after banner removal
+
 ## [1.1.2] - 2026-01-09
 
 ### Changed
@@ -62,6 +68,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Views Sheet Manager tools
 
 [1.1.2]: https://github.com/jason-svn/WWPTools/releases/tag/V1.1.2
+[1.1.3]: https://github.com/jason-svn/WWPTools/releases/tag/V1.1.3
 [1.1.1]: https://github.com/jason-svn/WWPTools/releases/tag/V1.1.1
 [1.1.0]: https://github.com/jason-svn/WWPTools/releases/tag/V1.1.0
 [1.0.0]: https://github.com/jason-svn/WWPTools/releases/tag/V1.0.0